Relize 1.0.1
Fixed a bug with hanging.
Now the device will reboot itself and start the clock in 10 seconds from the start.

The clock began to work faster. The interface has been rewritten to the lwjgl library

The Clock, Settings, Alarm Clock, can be started from the application menu.

Currently, the settings are started with a preload. That was the plan. In the future, this button will open applications (User programs, File Manager, Alarm Clock, Timer, Settings)